re-organizes code
